api: implement error handling for clCreateBuffer*
Also introduce a dedicated context header so we can use functions
from cvk_device in cvk_context.
